Random Product Generator :100:
The script has run successfully after I fixed the database (I had to remake the products table) :> the relationships were broken from my previous attempts to fix the database.

I currently have 205 items in the database which should be adequate for me to look into the layout of the genProducts page and add more details to the cards.

:> Future >>> Work on relationships and  pagination